Refactor: jpa query method 로 리팩토링
